# src/utils/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Swagger CRM Programming school",
    "description": "This is documentation for the CRM Programming school server. Here you will learn how to use this server.",
    "contact": {
      "email": "devalexhor@gmail.com"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/users/{userId}/": {
      "put": {
        "tags": ["users"],
        "summary": "Update one user by id",
        "operationId": "updateUser",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "description": "Access token",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user update",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "data",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the updated user.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "is_active": {
          "type": "boolean",
          "default": false
        },
        "is_superuser": {
          "type": "boolean",
          "default": false
        },
        "is_staff": {
          "type": "boolean",
          "default": false
        },
        "created_at": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        },
        "last_login": {
          "type": "string"
        }
      }
    },
...
  }
}